add compat file and added ensure_native_str to it.
